# docs/swagger.yaml
basePath: /api
definitions:
  auth.ForgotPasswordRequest:
    properties:
      email:
        type: string
    required:
    - email
    type: object
  auth.SignInRequest:
    properties:
      email:
        type: string
      password:
        type: string
    required:
    - email
    - password
    type: object
  auth.SignInResponse:
    properties:
      accessToken:
        type: string
      refreshToken:
        type: string
      tokenType:
        type: string
      user:
        $ref: '#/definitions/user.User'
    type: object
  auth.SignOutRequest:
    properties:
      refreshToken:
        type: string
    required:
    - refreshToken
    type: object
  auth.SignUpRequest:
    properties:
      email:
        type: string
      firstName:
        type: string
      lastName:
        type: string
      password:
        minLength: 6
        type: string
    required:
    - email
    - firstName
    - lastName
    - password
    type: object
  response.Response:
    properties:
      message:
        type: string
      payload: {}
      status:
        type: string
      success:
        type: boolean
    type: object
  todo.Todo:
    properties:
      completed:
        type: boolean
      createdAt:
        type: integer
      description:
        type: string
      id:
        type: integer
      title:
        type: string
      updatedAt:
        type: integer
    type: object
  todo.TodoCreateRequest:
    properties:
      description:
        type: string
      title:
        type: string
    required:
    - description
    - title
    type: object
  todo.TodoUpdateRequest:
    properties:
      completed:
        type: boolean
      description:
        type: string
      id:
        type: integer
      title:
        type: string
    required:
    - completed
    - description
    - id
    - title
    type: object
  user.UpdateRequest:
    properties:
      email:
        type: string
      firstName:
        type: string
      id:
        type: integer
      lastName:
        type: string
    required:
    - id
    type: object
  user.User:
    properties:
      createdAt:
        type: integer
      email:
        type: string
      firstName:
        type: string
      id:
        type: integer
      lastName:
        type: string
      updatedAt:
        type: integer
    type: object
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is a sample server Todos server.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/new
  title: Go Todos API
  version: "1.0"
paths:
  /auth/forgot-password:
    post:
      consumes:
      - application/json
      description: Send reset password email with token
      parameters:
      - description: Forgot Password Request
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/auth.ForgotPasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
      summary: Forgot Password
      tags:
      - auth-controller
  /auth/sign-in:
    post:
      consumes:
      - application/json
      description: Authenticate user and return JWT tokens
      parameters:
      - description: Sign In Request
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/auth.SignInRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/auth.SignInResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
      summary: User Sign In
      tags:
      - auth-controller
  /auth/sign-out:
    post:
      consumes:
      - application/json
      description: Invalidate refresh token and log user out
      parameters:
      - description: Sign Out Request
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/auth.SignOutRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
      summary: User Sign Out
      tags:
      - auth-controller
  /auth/sign-up:
    post:
      consumes:
      - application/json
      description: Create a new user account
      parameters:
      - description: Sign Up Request
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/auth.SignUpRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/user.User'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
      summary: User Sign Up
      tags:
      - auth-controller
  /todos:
    get:
      consumes:
      - application/json
      description: Get All Todos
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/todo.Todo'
            type: array
      security:
      - BearerAuth: []
      summary: Get All
      tags:
      - todo-controller
  /todos/{id}:
    get:
      consumes:
      - application/json
      description: Retrieves a todo item by its ID
      parameters:
      - description: Todo ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/todo.Todo'
        "400":
          description: Invalid ID
          schema:
            type: string
        "404":
          description: Todo not found
          schema:
            type: string
        "500":
          description: Internal server error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Get todo by ID
      tags:
      - todo-controller
  /todos/create:
    post:
      consumes:
      - application/json
      description: Creates a new todo item
      parameters:
      - description: create todo
        in: body
        name: todo
        required: true
        schema:
          $ref: '#/definitions/todo.TodoCreateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/todo.Todo'
        "400":
          description: Invalid request body
          schema:
            type: string
        "500":
          description: Internal server error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Create Todos Endpoint
      tags:
      - todo-controller
  /todos/delete/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a todo by ID
      parameters:
      - description: Todo ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Invalid ID
          schema:
            type: string
        "404":
          description: Todo not found
          schema:
            type: string
        "500":
          description: Internal server error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Delete a todo
      tags:
      - todo-controller
  /todos/update/{id}:
    put:
      consumes:
      - application/json
      description: Update todo details by ID
      parameters:
      - description: Todo ID
        in: path
        name: id
        required: true
        type: integer
      - description: Updated todo data
        in: body
        name: todo
        required: true
        schema:
          $ref: '#/definitions/todo.TodoUpdateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/todo.Todo'
        "400":
          description: Invalid input
          schema:
            type: string
        "404":
          description: Todo not found
          schema:
            type: string
        "500":
          description: Internal server error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Update a todo
      tags:
      - todo-controller
  /user/update/{id}:
    put:
      consumes:
      - application/json
      description: Update user details by ID
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      - description: Updated user data
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/user.UpdateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/user.User'
        "400":
          description: Invalid input
          schema:
            type: string
        "404":
          description: User not found
          schema:
            type: string
        "500":
          description: Internal server error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Update a user
      tags:
      - user-controller
  /users:
    get:
      consumes:
      - application/json
      description: Get All Users
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/user.User'
            type: array
      security:
      - BearerAuth: []
      summary: Get All
      tags:
      - user-controller
  /users/{id}:
    get:
      consumes:
      - application/json
      description: Retrieves a user by its ID
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/user.User'
        "400":
          description: Invalid ID
          schema:
            type: string
        "404":
          description: User not found
          schema:
            type: string
        "500":
          description: Internal server error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Get user by ID
      tags:
      - user-controller
  /users/delete/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a user by ID
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Invalid ID
          schema:
            type: string
        "404":
          description: User not found
          schema:
            type: string
        "500":
          description: Internal server error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Delete a user
      tags:
      - user-controller
  /users/me:
    get:
      consumes:
      - application/json
      description: Retrieves a user by its ID
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/user.User'
        "400":
          description: Invalid ID
          schema:
            type: string
        "404":
          description: User not found
          schema:
            type: string
        "500":
          description: Internal server error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Get current user
      tags:
      - user-controller
securityDefinitions:
  BearerAuth:
    description: Provide the access token directly, without "Bearer ".
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"